repair scale and rasize

+++ Content/DirectXTK3DSceneRenderer.h
@@ -69,10 +69,9 @@ namespace SimpleSample
 
         // Not usefull right now
 		//void XM_CALLCONV DrawGrid(DirectX::FXMVECTOR xAxis, DirectX::FXMVECTOR yAxis, DirectX::FXMVECTOR origin, size_t xdivs, size_t ydivs, DirectX::GXMVECTOR color);
-		
         std::unique_ptr<DirectX::SpriteBatch>                                   m_sprites;
         std::shared_ptr<DirectX::SpriteFont>                                    m_font;
-		int																		scale;
+		float																	scale;
 		//Sound
 		std::unique_ptr<DirectX::AudioEngine>                                   m_audEngine;
         std::unique_ptr<DirectX::WaveBank>                                      m_waveBank;
